* Add .npmrc file with FontAwesome token

* Update README with local setup instructions [sc-60] (#248)

* Update README with local setup instructions [sc-60]

* Add NPM steps to README instructions [sc-60]

* Authorize application actions using gate policy [sc-41]

* Authorize application staff comments using gate policy [sc-41]

* Change route URIs to comply with default resource structure [sc-71]

* Disable Trix preview captions and add mime type check on `trix-file-accept` [sc-422]

* Implement image uploading on event create page [sc-422]

* Implement image uploading on event edit page [sc-422]

* Re-enable Trix file button [sc-422]

* Fix profile role badges [sc-479] (#252)

* Merge Personal Job Overview & User Job Overview pages [sc-325] (#253)

* Merge Personal Job Overview User Job Overview pages [sc-325]

* Redirect personal job overview page to user job overview page [sc-325]

* Change personal overview link in navbar to user job overview page & Allow navbar dropdowns to pass parameters to the url [sc-325]

* Fix job overview page always returning authenticated user's balance [sc-325]

* Fix tables from overflowing off the page [sc-496] (#254)

* Fix table overflow on user profile page [sc-496]

* Fix table overflow on vacation request index page [sc-496]

* Fix table overflow on vacation request management page [sc-496]

* Fix table overflow on download management pages [sc-496]

* Fix table overflow on event management pages [sc-496]

+++ resources/views/livewire/recruitment/components/staff-comments.blade.php
@@ -24,13 +24,14 @@
                                         <p>{{ $comment->body }}</p>
                                     </div>
                                     <div class="mt-2 text-sm space-x-2">
-                                        <span
-                                            class="text-gray-500 font-medium">{{ \Carbon\Carbon::parse($comment->created_at)->diffForHumans() }}</span>
-                                        <span class="text-gray-500 font-medium">&middot;</span>
-                                        <button type="button" wire:click="deleteComment('{{ $comment->uuid }}')"
-                                                class="text-gray-900 font-medium">
-                                            Delete
-                                        </button>
+                                        <span class="text-gray-500 font-medium">{{ \Carbon\Carbon::parse($comment->created_at)->diffForHumans() }}</span>
+                                        @can('delete', $comment)
+                                            <span class="text-gray-500 font-medium">&middot;</span>
+                                            <button type="button" wire:click="deleteComment('{{ $comment->uuid }}')"
+                                                    class="text-gray-900 font-medium">
+                                                Delete
+                                            </button>
+                                        @endcan
                                     </div>
                                 </div>
                             </div>
